+++ search-films/src/components/LinkedMenuItem/index.jsx
@@ -1,4 +1,4 @@
-import { default as CssUtils } from '../shared/CssUtils';
+import cs from 'classnames';
 import styles from './index.module.css';
 
 /**
@@ -24,40 +24,34 @@ function LinkedMenuItem(props) {
 	const {canChoose=/*byDefault*/true} = props;
 	
 	const createIconItem = (icon) => {
-		const defaultIconClassName = styles['menu-item-icon'];
-
+		
 		if(icon && icon.type == 'svg') {
-			return <img className={defaultIconClassName} alt="Icon" src={icon.dataSource}/>;
+			return <img className={styles['menu-item-icon']} alt="Icon" src={icon.dataSource}/>;
 		}
 		if(icon && icon.type == 'iconText') {
-
-			const roundedIconClassName = CssUtils.addClassToDefaultClassName(
-				defaultIconClassName, 
-				styles['rounded-text-icon']);
-
-			return <span className={roundedIconClassName}>{icon.dataSource}</span>;
+			
+			return <span className={cs(
+				styles['menu-item-icon'], 
+				styles['rounded-text-icon'])}>{icon.dataSource}</span>;
 		}
 		/**wrong format of icon object or icon not present */
 		return;
 	};
 
 	const iconItem = createIconItem(props.icon);
-	const defaultMenuItemClassName = styles['menu-item'];
-	const menuItemClassName = canChoose ? (
-	/*byDefault*/ CssUtils.addClassToDefaultClassName(defaultMenuItemClassName, styles['can-choose'])) : defaultMenuItemClassName;
+
 	const onMenuClick = (e) =>{
 		const handleEnabled = props.onClick && canChoose;
 		if(handleEnabled) {
 			props.onClick(e);
 		}
 	};
-	const defaultLinkClassName = styles['menu-link'];
-	const disabledLink = CssUtils.addClassToDefaultClassName(defaultLinkClassName, styles['disable']);
-	const linkClassName = canChoose ? /*enable link byDefault*/ defaultLinkClassName : disabledLink;
+
 	return (
-		<li className={menuItemClassName} onClick={onMenuClick}>
-			<a href={props.linkUrl} className={linkClassName}>{props.text} {iconItem}</a>
-			
+		<li className={cs(styles['menu-item'], {[styles['can-choose']]: canChoose})} onClick={onMenuClick}>
+			<a href={props.linkUrl} className={cs(styles['menu-link'], {[styles['disable']]: canChoose == false})}>
+				{props.text} {iconItem}
+			</a>
 		</li>
 	);
 }